+++ NEWS.md
@@ -13,6 +13,8 @@
 
 * The default value of `failure` argument in `parse_http_date()` is set to `structure(NA_real_, class = "Date")` so that the reponse with a "failure" date can be printed out correctly. (@shrektan, #544)
 
+* Non-http(s) methods now give an informative warning that headers are not interpretable the same way, and header parsing only minimally occurs. (@billdenney, #537)
+
 # httr 1.3.1
 
 * Re-enable on-disk caching (accidentally disabled in #457) (#475)